for both the layouts wrappers are added , as the server fuctions can be applied in layout files
