+++ src/app/dashboard/hats/page.js
