+++ src/styles/globals.css
@@ -1,0 +1,87 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+:root {
+  --foreground-rgb: 0, 0, 0;
+  --background-start-rgb: 214, 219, 220;
+  --background-end-rgb: 255, 255, 255;
+}
+
+@media (prefers-color-scheme: dark) {
+  :root {
+    --foreground-rgb: 255, 255, 255;
+    --background-start-rgb: 0, 0, 0;
+    --background-end-rgb: 0, 0, 0;
+  }
+}
+
+* {
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+}
+
+html,
+body {
+  max-width: 100vw;
+  overflow-x: hidden;
+}
+
+@layer utilities {
+  .text-balance {
+    text-wrap: balance;
+  }
+}
+
+
+.full-screen {
+  width: 100vw;
+  height: 100vh;
+  overflow: hidden;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+
+/* subsciption page */
+
+.no-after-absolute::after {
+  position: static !important;
+  /* Override absolute positioning */
+}
+
+/* documents */
+
+.document-background-image {
+  position: fixed !important;
+  height: 75% !important;
+  width: 35% !important;
+  min-width: 297px;
+  align-self: center;
+  left: 0;
+  top: 2.5%;
+  right: 0;
+  bottom: 0;
+  color: transparent;
+  z-index: -1;
+  opacity: 0.7;
+}
+
+/* auth */
+
+.auth-background-image {
+  position: fixed !important;
+  height: 75% !important;
+  width: 35% !important;
+  min-width: 297px;
+  align-self: center;
+  left: 0;
+  top: 2.5%;
+  right: 0;
+  bottom: 0;
+  color: transparent;
+  z-index: -1;
+  opacity: 1;
+}
